refactor: 💡 use specific files for classes
Instead of putting everything on the same file, split it to make it more
readable
